Add SOURCEDIRS and discovery of .c files
It tends to be easy to miss including .c-files, but directories is
harder to miss. Especially when combining with modular includes, so
modules are just added to sections as list.

If missing including a .c file will silently pass the build, and
therefore get no feedback if the a .c file is missing until it is
used at runtime