Add systimer + runtime analysis
Monitoring CPU usage is crucial to build a stable system. This will
monitor the CPU usage for each task over time.

It would be helpful to get a 10 second average to see how it changes,
but at least a first step.

It reserves TIM7, which is rarely useful for PWM anyway, to make sure
to have access to the timer from scheduler start